Change display of page for leaving draw
This previously consisted of a table of Members with the option to remove
each one. The content has been changed to show the option to remove the
current user at the top, followed by a table to remove other users.
Tests have been rewritten and refactored to reflect the changes. The
flash for leaving the draw has been changed to show different content
depending on whether the user is removing themself or another user from
the draw.